requirements.txt
Django~=4.2.15
djangorestframework~=3.15.2
django-ckeditor~=6.7.1
h11~=0.14.0
pillow~=10.4.0
pip~=22.0.4
tornado~=6.4.1
pytz~=2024.1
exceptiongroup~=1.2.2
sniffio~=1.3.1
httpcore~=1.0.5
idna~=3.8
certifi~=2024.8.30
asgiref~=3.8.1
sqlparse~=0.5.1
tzlocal~=5.2
anyio~=4.4.0
APScheduler~=3.6.3
six~=1.16.0
setuptools~=58.1.0